Update module mvp package name
