+++ client/src/main/java/engineer/mathsoftware/cp/dtfs/client/App.java
@@ -18,9 +18,9 @@ import engineer.mathsoftware.cp.dtfs.client.header.Header;
 import engineer.mathsoftware.cp.dtfs.client.info.Info;
 import engineer.mathsoftware.cp.dtfs.client.menu.AppMenu;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.file.text.TextFileRepository;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.Initializable;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpPresenter;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpView;
+import engineer.mathsoftware.cp.dtfs.mvp.Initializable;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpPresenter;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpView;
 import com.github.tobiasbriones.cp.rmifilesystem.model.FileSystemService;
 import javafx.application.Platform;
 import javafx.scene.Node;
