+++ client/src/main/java/engineer/mathsoftware/cp/dtfs/client/AppPresenter.java
@@ -13,7 +13,7 @@
 
 package engineer.mathsoftware.cp.dtfs.client;
 
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.AbstractMvpPresenter;
+import engineer.mathsoftware.cp.dtfs.mvp.AbstractMvpPresenter;
 import engineer.mathsoftware.cp.dtfs.client.header.Header;
 
 /**
