+++ client/src/main/java/engineer/mathsoftware/cp/dtfs/client/content/editor/Editor.java
@@ -15,9 +15,9 @@ package engineer.mathsoftware.cp.dtfs.client.content.editor;
 
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.File;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.file.text.TextFileRepository;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.Initializable;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpPresenter;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpView;
+import engineer.mathsoftware.cp.dtfs.mvp.Initializable;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpPresenter;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpView;
 import javafx.scene.Node;
 
 /**
