+++ client/src/main/java/engineer/mathsoftware/cp/dtfs/client/content/files/Files.java
@@ -18,9 +18,9 @@ import com.github.tobiasbriones.cp.rmifilesystem.model.io.File;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.file.text.TextFileRepository;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.node.DirectoryNode;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.node.FileSystem;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.Initializable;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpPresenter;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpView;
+import engineer.mathsoftware.cp.dtfs.mvp.Initializable;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpPresenter;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpView;
 import com.github.tobiasbriones.cp.rmifilesystem.model.io.node.Node;
 
 /**
