+++ client/src/main/java/engineer/mathsoftware/cp/dtfs/client/header/Header.java
@@ -13,9 +13,9 @@
 
 package engineer.mathsoftware.cp.dtfs.client.header;
 
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.Initializable;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpPresenter;
-import com.github.tobiasbriones.cp.rmifilesystem.mvp.MvpView;
+import engineer.mathsoftware.cp.dtfs.mvp.Initializable;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpPresenter;
+import engineer.mathsoftware.cp.dtfs.mvp.MvpView;
 import javafx.scene.Node;
 
 /**
