+++ mvp/src/main/java/engineer/mathsoftware/cp/dtfs/mvp/Initializable.java
@@ -11,26 +11,12 @@
  * https://opensource.org/licenses/BSD-3-Clause.
  */
 
-package com.github.tobiasbriones.cp.rmifilesystem.mvp;
-
-import java.util.Optional;
+package engineer.mathsoftware.cp.dtfs.mvp;
 
 /**
  * @author Tobias Briones
  */
-public abstract class AbstractMvpPresenter<O> implements MvpPresenter<O> {
-    private O output;
-
-    protected AbstractMvpPresenter() {
-        output = null;
-    }
-
-    protected Optional<O> getOutput() {
-        return Optional.ofNullable(output);
-    }
-
-    @Override
-    public final void setOutput(O value) {
-        output = value;
-    }
+@FunctionalInterface
+public interface Initializable {
+    void init();
 }
